Bloom effect looks even better when number of mipmaps is dialed up.

+++ src/graphics/screen/screenManager.cpp
@@ -63,7 +63,7 @@ void ScreenManager::generateFrameBuffers() {
 
   glGenTextures(1, &m_screenHDRTexture);
   glBindTexture(GL_TEXTURE_2D, m_screenHDRTexture);
-  glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
+  glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
   glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
   glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
   glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
@@ -83,7 +83,7 @@ void ScreenManager::generateFrameBuffers() {
   glGenTextures(m_screenBloomTextures.size(), &m_screenBloomTextures[0]);
   for (unsigned int i = 0; i < m_screenBloomTextures.size(); i++) {
     glBindTexture(GL_TEXTURE_2D, m_screenBloomTextures[i]);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width >> (i), height >> (i), 0, GL_RGBA, GL_FLOAT, NULL);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, width >> (i), height >> (i), 0, GL_RGBA, GL_FLOAT, NULL);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
@@ -112,7 +112,7 @@ float ScreenManager::calculateLuminance() {
   int rangeX = width * range;
   int rangeY = height * range;
 
-  glBindImageTexture(2, m_screenHDRTexture, 0, GL_FALSE, 0, GL_READ_ONLY, GL_RGBA16F);
+  glBindImageTexture(2, m_screenHDRTexture, 0, GL_FALSE, 0, GL_READ_ONLY, GL_RGBA32F);
 
   ShaderManager* shaderManager = ShaderManager::getInstance();
   shaderManager->bindComputeShader("../assets/shaders/compute/calculateLuminance.comp");
@@ -184,7 +184,6 @@ void ScreenManager::applyBloom() {
   // Downsample hdr buffer into its mipmaps
   shaderManager->bindComputeShader("../assets/shaders/compute/bloom_downsample.comp");
   glActiveTexture(GL_TEXTURE0);
-
    
   // Copy hdrRender into bloom level 0
   glCopyImageSubData(m_screenHDRTexture, GL_TEXTURE_2D, 0, 0, 0, 0, 
@@ -198,7 +197,7 @@ void ScreenManager::applyBloom() {
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-    glBindImageTexture(3, m_screenBloomTextures[mipMapLevel], 0, GL_FALSE, 0, GL_WRITE_ONLY, GL_RGBA16F);
+    glBindImageTexture(3, m_screenBloomTextures[mipMapLevel], 0, GL_FALSE, 0, GL_WRITE_ONLY, GL_RGBA32F);
 
     const unsigned int adjustedGroupSize_X = std::ceil( (width >> (mipMapLevel)) / 32.0f);
     const unsigned int adjustedGroupSize_Y = std::ceil( (height >> (mipMapLevel)) / 32.0f);
@@ -218,7 +217,7 @@ void ScreenManager::applyBloom() {
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-    glBindImageTexture(3, m_screenBloomTextures[mipMapLevel-1], 0, GL_FALSE, 0, GL_READ_WRITE, GL_RGBA16F);
+    glBindImageTexture(3, m_screenBloomTextures[mipMapLevel-1], 0, GL_FALSE, 0, GL_READ_WRITE, GL_RGBA32F);
 
     const unsigned int adjustedGroupSize_X = std::ceil((width >> (mipMapLevel-1)) / 32.0f);
     const unsigned int adjustedGroupSize_Y = std::ceil((height >> (mipMapLevel-1)) / 32.0f);
